use `double` (not `float`) for quantities.
while float will have more than adequate precision, using doubles makes
the Java side of things rather more graceful (no literal suffix). those
extra bytes of storage are "whatever", at least for now.

+++ src/main/java/com/brennaswitzer/cookbook/domain/Recipe.java
@@ -72,7 +72,7 @@ public class Recipe extends Ingredient implements AggregateIngredient, Owned {
     }
 
     @Override
-    public void addIngredient(Float quantity, String units, Ingredient ingredient, String preparation) {
+    public void addIngredient(Double quantity, String units, Ingredient ingredient, String preparation) {
         ensureIngredients();
         ingredients.add(new IngredientRef<>(quantity, units, ingredient, preparation));
     }
